Funciones Array hasta DesviaciónTipica
